# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenRisk API
  version: 1.0.0
  description: |
    Modern, intelligent risk management platform API.
    Part of the OpenDefender suite with AI-driven risk assessment, mitigation tracking, and security integration.
    
    **Authentication**: All protected endpoints require a Bearer JWT token in the `Authorization` header.
    
    **Base URL**: `/api/v1`
  contact:
    name: OpenDefender Team
  license:
    name: Business Source License 1.1 (BUSL-1.1)
    url: https://mariadb.com/bsl11/
servers:
  - url: /api/v1
    description: Production API
...
paths:
...
  /risks:
    get:
      tags:
        - Risks
      summary: List all risks
      operationId: listRisks
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: sort_by
          in: query
          schema:
            type: string
            description: Field to sort by (e.g., score, title, impact)
      responses:
        '200':
          description: List of risks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Risk'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Risks
      summary: Create a new risk
      operationId: createRisk
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRiskInput'
      responses:
        '201':
          description: Risk created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /risks/{id}:
    get:
      tags:
        - Risks
      summary: Get risk details
      operationId: getRisk
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Risk UUID
      responses:
        '200':
          description: Risk details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '401':
          description: Unauthorized
        '404':
          description: Risk not found

    patch:
      tags:
        - Risks
      summary: Update risk
      operationId: updateRisk
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRiskInput'
      responses:
        '200':
          description: Risk updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Risk not found

    delete:
      tags:
        - Risks
      summary: Delete risk
      operationId: deleteRisk
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Risk deleted
        '401':
          description: Unauthorized
        '404':
          description: Risk not found

  # ==================== MITIGATIONS ====================
...
  /stats/trends:
    get:
      tags:
        - Statistics
      summary: Get global risk trend
      operationId: getRiskTrend
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Risk trend data
          content:
            application/json:
              schema:
                type: object

  # ==================== EXPORT ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token. Obtain via /auth/login
  schemas:
...
    CreateRiskInput:
      type: object
      required:
        - title
        - impact
        - probability
      properties:
        title:
          type: string
          minLength: 3
          example: "Data breach risk"
        description:
          type: string
          example: "Unauthorized access to sensitive customer data"
        impact:
          type: integer
          minimum: 1
          maximum: 5
          description: Impact level (1-5)
          example: 5
        probability:
          type: integer
          minimum: 1
          maximum: 5
          description: Probability level (1-5)
          example: 3
        tags:
          type: array
          items:
            type: string
          example: ["security", "data-protection"]
        asset_ids:
          type: array
          items:
            type: string
            format: uuid
        framework:
          type: string
          enum: [ISO27001, CIS, NIST, OWASP]
          example: ISO27001
    UpdateRiskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        impact:
          type: integer
          minimum: 1
          maximum: 5
        probability:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          items:
            type: string
        asset_ids:
          type: array
          items:
            type: string
            format: uuid
...
    Risk:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        impact:
          type: integer
          minimum: 1
          maximum: 5
        probability:
          type: integer
          minimum: 1
          maximum: 5
        score:
          type: number
          format: float
          description: Calculated risk score
        status:
          type: string
          enum: [OPEN, MITIGATED, ACCEPTED, CLOSED]
        framework:
          type: string
          enum: [ISO27001, CIS, NIST, OWASP]
        tags:
          type: array
          items:
            type: string
        assets:
          type: array
          items:
            $ref: '#/components/schemas/Asset'
        mitigations:
          type: array
          items:
            $ref: '#/components/schemas/Mitigation'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Mitigation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        risk_id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [PLANNED, IN_PROGRESS, DONE]
        assignee:
          type: string
        progress:
          type: integer
          minimum: 0
          maximum: 100
        cost:
          type: string
        mitigation_time:
          type: string
        due_date:
          type: string
          format: date
        sub_actions:
          type: array
          items:
            $ref: '#/components/schemas/MitigationSubAction'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    MitigationSubAction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mitigation_id:
          type: string
          format: uuid
        title:
          type: string
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [Server, Laptop, Database, SaaS, Network, Storage]
        criticality:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        owner:
          type: string
        source:
          type: string
          enum: [MANUAL, OPENASSET]
        external_id:
          type: string
        risks:
          type: array
          items:
            $ref: '#/components/schemas/Risk'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid input"
        code:
          type: integer
          example: 400
        details:
          type: object